Remove prop-drilling from PriceCalculator sections and fields
- Read values directly from specific fields instead of prop-drilling via PriceCalculatorAccordion and AutomaticField
- Remove PriceCalculatorAccordion, it had too much coupling to PriceCalculator
- Stop using render child pattens since now we don't need to prop-drill so much
